# server/src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RunMe API",
    "description": "REST API for the RunMe mobile app (auth, loans, grants, wallet, admin). Use **Authorize** with a Bearer token from `POST /auth/login` or `POST /auth/register`.",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "/", "description": "Current host" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the `accessToken` value from login/register (without the word Bearer)."
      }
    },
    "schemas": {
...
      "Grant": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "story": { "type": "string" },
          "goalNaira": { "type": "number" },
          "raisedNaira": { "type": "number" },
          "category": { "type": "string", "enum": ["studentStory", "urgentNeed", "education"] },
          "studentName": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "attachmentUrls": { "type": "array", "items": { "type": "string" } },
          "isUrgent": { "type": "boolean" }
        }
      },
      "GrantCreate": {
        "type": "object",
        "required": ["title", "story", "goalNaira", "category", "studentName"],
        "properties": {
          "title": { "type": "string" },
          "story": { "type": "string" },
          "goalNaira": { "type": "number", "minimum": 0 },
          "category": { "type": "string", "enum": ["studentStory", "urgentNeed", "education"] },
          "studentName": { "type": "string" },
          "isUrgent": { "type": "boolean" },
          "attachmentUrls": { "type": "array", "items": { "type": "string", "format": "uri" } }
        }
      },
      "DonationCreate": {
        "type": "object",
        "required": ["amountNaira"],
        "properties": {
          "amountNaira": { "type": "number", "minimum": 0 },
          "anonymous": { "type": "boolean" }
        }
      },
      "CommentCreate": {
        "type": "object",
        "required": ["authorName", "message"],
        "properties": {
          "authorName": { "type": "string" },
          "message": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/admin/grants": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all grants",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
    "/grants": {
      "get": {
        "tags": ["Grants"],
        "summary": "List grants",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string", "enum": ["studentStory", "urgentNeed", "education"] } },
          { "name": "urgentOnly", "in": "query", "schema": { "type": "boolean" } },
          { "name": "excludeBorrowerId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/Grant" } },
                    "nextCursor": { "type": "string", "nullable": true }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Grants"],
        "summary": "Create grant request",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/GrantCreate" } }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Grant" } } }
          }
        }
      }
    },
    "/grants/{grantId}": {
      "get": {
        "tags": ["Grants"],
        "summary": "Get grant",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "grantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Grant" } } }
          },
          "404": { "description": "Not found" }
        }
      }
    },
    "/grants/{grantId}/donations": {
      "get": {
        "tags": ["Grants"],
        "summary": "List donations",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "grantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not found" }
        }
      },
      "post": {
        "tags": ["Grants"],
        "summary": "Donate",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "grantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/DonationCreate" } }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/grants/{grantId}/comments": {
      "get": {
        "tags": ["Grants"],
        "summary": "List comments",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "grantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not found" }
        }
      },
      "post": {
        "tags": ["Grants"],
        "summary": "Post comment",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "grantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CommentCreate" } }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "404": { "description": "Not found" }
        }
      }
    },
...
  }
}